@@ -1,19 +1,19 @@
 name: device_info_plus_windows
 description: Windows implementation of the device_info_plus plugin
-version: 2.2.0
+version: 3.0.2
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
   device_info_plus_platform_interface: ^2.1.0
-  ffi: ^1.0.0
+  ffi: ^2.0.1
   flutter:
     sdk: flutter
-  win32: ^2.0.0
+  win32: ^2.7.0
 
 dev_dependencies:
   flutter_test:
