@@ -34,8 +34,8 @@ class ShareWindows extends SharePlatform {
           .join('&'),
     );
 
-    if (await canLaunch(uri.toString())) {
-      await launch(uri.toString());
+    if (await canLaunchUrl(uri)) {
+      await launchUrl(uri);
     } else {
       throw Exception('Unable to share on windows');
     }
