@@ -202,8 +202,8 @@ public class AlarmService extends JobIntentService {
         context,
         request.requestCode,
         alarmClock,
-        repeating,
         request.allowWhileIdle,
+        repeating,
         request.exact,
         request.wakeup,
         request.startMillis,
