@@ -1,11 +1,11 @@
 import 'dart:io';
 
-import 'package:android_intent_plus/android_intent.dart';
 import 'package:android_intent_example/main.dart';
-import 'package:integration_test/integration_test.dart';
+import 'package:android_intent_plus/android_intent.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:integration_test/integration_test.dart';
 
 /// This is a smoke test that verifies that the example app builds and loads.
 /// Because this plugin works by launching Android platform UIs it's not
@@ -13,6 +13,7 @@ import 'package:flutter_test/flutter_test.dart';
 /// are more useful unit tests for the platform logic under android/src/test/.
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+
   testWidgets('Embedding example app loads', (WidgetTester tester) async {
     // Build our app and trigger a frame.
     await tester.pumpWidget(const MyApp());
