@@ -2,17 +2,16 @@ import 'dart:async';
 
 import 'package:battery_plus_platform_interface/battery_plus_platform_interface.dart';
 import 'package:meta/meta.dart';
+import 'package:upower/upower.dart';
 
-import 'upower_device.dart';
-
-extension _ToBatteryState on UPowerBatteryState {
+extension _ToBatteryState on UPowerDeviceState {
   BatteryState toBatteryState() {
     switch (this) {
-      case UPowerBatteryState.charging:
+      case UPowerDeviceState.charging:
         return BatteryState.charging;
-      case UPowerBatteryState.discharging:
+      case UPowerDeviceState.discharging:
         return BatteryState.discharging;
-      case UPowerBatteryState.fullyCharged:
+      case UPowerDeviceState.fullyCharged:
         return BatteryState.full;
       default:
         return BatteryState.unknown;
@@ -22,7 +21,7 @@ extension _ToBatteryState on UPowerBatteryState {
 
 ///
 @visibleForTesting
-typedef UPowerDeviceFactory = UPowerDevice Function();
+typedef UPowerClientFactory = UPowerClient Function();
 
 /// The Linux implementation of BatteryPlatform.
 class BatteryPlusLinux extends BatteryPlatform {
@@ -34,21 +33,21 @@ class BatteryPlusLinux extends BatteryPlatform {
   /// Returns the current battery level in percent.
   @override
   Future<int> get batteryLevel {
-    final device = createDevice();
-    return device
-        .getPercentage()
-        .then((value) => value.round())
-        .whenComplete(() => device.dispose());
+    final client = createClient();
+    return client
+        .connect()
+        .then((_) => client.displayDevice.percentage.round())
+        .whenComplete(() => client.close());
   }
 
   /// Returns the current battery state.
   @override
   Future<BatteryState> get batteryState {
-    final device = createDevice();
-    return device
-        .getState()
-        .then((value) => value.toBatteryState())
-        .whenComplete(() => device.dispose());
+    final client = createClient();
+    return client
+        .connect()
+        .then((_) => client.displayDevice.state.toBatteryState())
+        .whenComplete(() => client.close);
   }
 
   /// Fires whenever the battery state changes.
@@ -61,25 +60,31 @@ class BatteryPlusLinux extends BatteryPlatform {
     return _stateController!.stream.asBroadcastStream();
   }
 
-  UPowerDevice? _stateDevice;
+  UPowerClient? _stateClient;
   StreamController<BatteryState>? _stateController;
 
   @visibleForTesting
   // ignore: public_member_api_docs, prefer_function_declarations_over_variables
-  UPowerDeviceFactory createDevice = () => UPowerDevice.display();
+  UPowerClientFactory createClient = () => UPowerClient();
 
-  void _addState(UPowerBatteryState value) {
+  void _addState(UPowerDeviceState value) {
     _stateController!.add(value.toBatteryState());
   }
 
-  void _startListenState() {
-    _stateDevice ??= createDevice();
-    _stateDevice!.getState().then((value) => _addState(value));
-    _stateDevice!.subscribeStateChanged().listen((value) => _addState(value));
+  Future<void> _startListenState() async {
+    _stateClient ??= createClient();
+    await _stateClient!
+        .connect()
+        .then((_) => _addState(_stateClient!.displayDevice.state));
+    _stateClient!.displayDevice.propertiesChanged.listen((properties) {
+      if (properties.contains('State')) {
+        _addState(_stateClient!.displayDevice.state);
+      }
+    });
   }
 
   void _stopListenState() {
-    _stateDevice?.dispose();
-    _stateDevice = null;
+    _stateClient?.close();
+    _stateClient = null;
   }
 }
