@@ -1,13 +1,13 @@
 import 'package:battery_plus_linux/src/battery_plus_linux.dart';
-import 'package:battery_plus_linux/src/upower_device.dart';
 import 'package:battery_plus_platform_interface/battery_plus_platform_interface.dart';
-import 'package:dbus/src/dbus_client.dart';
-import 'package:dbus/src/dbus_introspect.dart';
-import 'package:dbus/src/dbus_method_response.dart';
-import 'package:dbus/src/dbus_remote_object.dart';
-import 'package:dbus/src/dbus_value.dart';
+import 'package:mockito/annotations.dart';
+import 'package:mockito/mockito.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:upower/upower.dart';
 
+import 'battery_plus_linux_test.mocks.dart';
+
+@GenerateMocks([UPowerClient, UPowerDevice])
 void main() {
   test('registered instance', () {
     BatteryPlusLinux.registerWith();
@@ -15,102 +15,49 @@ void main() {
   });
   test('battery level', () async {
     final battery = BatteryPlusLinux();
-    battery.createDevice = () {
-      return MockDevice();
+    battery.createClient = () {
+      return createMockClient(percentage: 56.78);
     };
     expect(battery.batteryLevel, completion(equals(57)));
   });
 
   test('battery state', () async {
     final battery = BatteryPlusLinux();
-    battery.createDevice = () {
-      return MockDevice();
+    battery.createClient = () {
+      return createMockClient(state: UPowerDeviceState.charging);
     };
     expect(battery.batteryState, completion(BatteryState.charging));
   });
 
   test('battery state changes', () {
     final battery = BatteryPlusLinux();
-    battery.createDevice = () {
-      return MockDevice();
+    battery.createClient = () {
+      final client = createMockClient(state: UPowerDeviceState.charging);
+      final device = client.displayDevice;
+      when(device.propertiesChanged).thenAnswer((_) {
+        when(device.state).thenReturn(UPowerDeviceState.fullyCharged);
+        return Stream.value(['State']);
+      });
+      return client;
     };
     expect(battery.onBatteryStateChanged,
         emitsInOrder([BatteryState.charging, BatteryState.full]));
   });
 }
 
-class MockDevice implements UPowerDevice {
-  @override
-  Future<DBusMethodSuccessResponse> callMethod(
-    String? interface,
-    String member,
-    Iterable<DBusValue> values, {
-    DBusSignature? replySignature,
-    bool noReplyExpected = false,
-    bool noAutoStart = false,
-    bool allowInteractiveAuthorization = false,
-  }) async {
-    throw UnimplementedError();
+MockUPowerClient createMockClient({
+  double? percentage,
+  UPowerDeviceState? state,
+}) {
+  final device = MockUPowerDevice();
+  if (percentage != null) {
+    when(device.percentage).thenReturn(percentage);
+  }
+  if (state != null) {
+    when(device.state).thenReturn(state);
   }
 
-  @override
-  DBusClient get client => throw UnimplementedError();
-
-  @override
-  void dispose() {}
-
-  @override
-  Future<Map<String, DBusValue>> getAllProperties(String interface) {
-    throw UnimplementedError();
-  }
-
-  @override
-  Future<double> getPercentage() {
-    return Future.value(56.78);
-  }
-
-  @override
-  Future<DBusValue> getProperty(
-    String interface,
-    String name, {
-    DBusSignature? signature,
-  }) {
-    throw UnimplementedError();
-  }
-
-  @override
-  Future<UPowerBatteryState> getState() {
-    return Future.value(UPowerBatteryState.charging);
-  }
-
-  @override
-  Future<DBusIntrospectNode> introspect() {
-    throw UnimplementedError();
-  }
-
-  @override
-  String get name => throw UnimplementedError();
-
-  @override
-  DBusObjectPath get path => throw UnimplementedError();
-
-  @override
-  Future<void> setProperty(String interface, String name, DBusValue value) {
-    throw UnimplementedError();
-  }
-
-  @override
-  Stream<DBusPropertiesChangedSignal> get propertiesChanged {
-    throw UnimplementedError();
-  }
-
-  @override
-  set propertiesChanged(Stream<DBusPropertiesChangedSignal> propertiesChanged) {
-    throw UnimplementedError();
-  }
-
-  @override
-  Stream<UPowerBatteryState> subscribeStateChanged() {
-    return Stream.value(UPowerBatteryState.fullyCharged);
-  }
+  final client = MockUPowerClient();
+  when(client.displayDevice).thenReturn(device);
+  return client;
 }
