@@ -1,26 +1,30 @@
-// Mocks generated by Mockito 5.0.7 from annotations
+// Mocks generated by Mockito 5.2.0 from annotations
 // in connectivity_plus_linux/test/connectivity_plus_linux_test.dart.
 // Do not manually edit this file.
 
 import 'dart:async' as _i3;
 
+import 'package:dbus/dbus.dart' as _i4;
 import 'package:mockito/mockito.dart' as _i1;
 import 'package:nm/src/network_manager_client.dart' as _i2;
 
-// ignore_for_file: comment_references
-// ignore_for_file: unnecessary_parenthesis
-
-// ignore_for_file: prefer_const_constructors
-
+// ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
+// ignore_for_file: avoid_setters_without_getters
+// ignore_for_file: comment_references
+// ignore_for_file: implementation_imports
+// ignore_for_file: invalid_use_of_visible_for_testing_member
+// ignore_for_file: prefer_const_constructors
+// ignore_for_file: unnecessary_parenthesis
+// ignore_for_file: camel_case_types
 
-class _FakeNetworkManagerSettings extends _i1.Fake
+class _FakeNetworkManagerSettings_0 extends _i1.Fake
     implements _i2.NetworkManagerSettings {}
 
-class _FakeNetworkManagerDnsManager extends _i1.Fake
+class _FakeNetworkManagerDnsManager_1 extends _i1.Fake
     implements _i2.NetworkManagerDnsManager {}
 
-class _FakeNetworkManagerActiveConnection extends _i1.Fake
+class _FakeNetworkManagerActiveConnection_2 extends _i1.Fake
     implements _i2.NetworkManagerActiveConnection {}
 
 /// A class which mocks [NetworkManagerClient].
@@ -126,35 +130,53 @@ class MockNetworkManagerClient extends _i1.Mock
       (super.noSuchMethod(Invocation.getter(#connectivityCheckUri),
           returnValue: '') as String);
   @override
+  _i2.NetworkManagerState get state => (super.noSuchMethod(
+      Invocation.getter(#state),
+      returnValue: _i2.NetworkManagerState.unknown) as _i2.NetworkManagerState);
+  @override
   _i2.NetworkManagerSettings get settings =>
       (super.noSuchMethod(Invocation.getter(#settings),
-              returnValue: _FakeNetworkManagerSettings())
+              returnValue: _FakeNetworkManagerSettings_0())
           as _i2.NetworkManagerSettings);
   @override
   _i2.NetworkManagerDnsManager get dnsManager =>
       (super.noSuchMethod(Invocation.getter(#dnsManager),
-              returnValue: _FakeNetworkManagerDnsManager())
+              returnValue: _FakeNetworkManagerDnsManager_1())
           as _i2.NetworkManagerDnsManager);
   @override
   _i3.Future<void> connect() =>
       (super.noSuchMethod(Invocation.method(#connect, []),
-          returnValue: Future<void>.value(null),
-          returnValueForMissingStub: Future.value()) as _i3.Future<void>);
+          returnValue: Future<void>.value(),
+          returnValueForMissingStub: Future<void>.value()) as _i3.Future<void>);
   @override
   _i3.Future<void> setWirelessEnabled(bool? value) =>
       (super.noSuchMethod(Invocation.method(#setWirelessEnabled, [value]),
-          returnValue: Future<void>.value(null),
-          returnValueForMissingStub: Future.value()) as _i3.Future<void>);
+          returnValue: Future<void>.value(),
+          returnValueForMissingStub: Future<void>.value()) as _i3.Future<void>);
   @override
   _i3.Future<void> setWwanEnabled(bool? value) =>
       (super.noSuchMethod(Invocation.method(#setWwanEnabled, [value]),
-          returnValue: Future<void>.value(null),
-          returnValueForMissingStub: Future.value()) as _i3.Future<void>);
+          returnValue: Future<void>.value(),
+          returnValueForMissingStub: Future<void>.value()) as _i3.Future<void>);
   @override
   _i3.Future<void> setConnectivityCheckEnabled(bool? value) => (super
       .noSuchMethod(Invocation.method(#setConnectivityCheckEnabled, [value]),
-          returnValue: Future<void>.value(null),
-          returnValueForMissingStub: Future.value()) as _i3.Future<void>);
+          returnValue: Future<void>.value(),
+          returnValueForMissingStub: Future<void>.value()) as _i3.Future<void>);
+  @override
+  _i3.Future<_i2.NetworkManagerActiveConnection> addAndActivateConnection(
+          {Map<String, Map<String, _i4.DBusValue>>? connection = const {},
+          _i2.NetworkManagerDevice? device,
+          _i2.NetworkManagerAccessPoint? accessPoint}) =>
+      (super.noSuchMethod(
+              Invocation.method(#addAndActivateConnection, [], {
+                #connection: connection,
+                #device: device,
+                #accessPoint: accessPoint
+              }),
+              returnValue: Future<_i2.NetworkManagerActiveConnection>.value(
+                  _FakeNetworkManagerActiveConnection_2()))
+          as _i3.Future<_i2.NetworkManagerActiveConnection>);
   @override
   _i3.Future<_i2.NetworkManagerActiveConnection> activateConnection(
           {_i2.NetworkManagerDevice? device,
@@ -167,17 +189,17 @@ class MockNetworkManagerClient extends _i1.Mock
                 #accessPoint: accessPoint
               }),
               returnValue: Future<_i2.NetworkManagerActiveConnection>.value(
-                  _FakeNetworkManagerActiveConnection()))
+                  _FakeNetworkManagerActiveConnection_2()))
           as _i3.Future<_i2.NetworkManagerActiveConnection>);
   @override
   _i3.Future<void> deactivateConnection(
           _i2.NetworkManagerActiveConnection? connection) =>
       (super.noSuchMethod(
           Invocation.method(#deactivateConnection, [connection]),
-          returnValue: Future<void>.value(null),
-          returnValueForMissingStub: Future.value()) as _i3.Future<void>);
+          returnValue: Future<void>.value(),
+          returnValueForMissingStub: Future<void>.value()) as _i3.Future<void>);
   @override
   _i3.Future<void> close() => (super.noSuchMethod(Invocation.method(#close, []),
-      returnValue: Future<void>.value(null),
-      returnValueForMissingStub: Future.value()) as _i3.Future<void>);
+      returnValue: Future<void>.value(),
+      returnValueForMissingStub: Future<void>.value()) as _i3.Future<void>);
 }
