@@ -4,6 +4,7 @@
 
 #import "FLTSharePlusPlugin.h"
 #import "LinkPresentation/LPLinkMetadata.h"
+#import "LinkPresentation/LPMetadataProvider.h"
 
 static NSString *const PLATFORM_CHANNEL = @"dev.fluttercommunity.plus/share";
 
@@ -75,8 +76,13 @@ TopViewControllerForViewController(UIViewController *viewController) {
 
 - (instancetype)initWithSubject:(NSString *)subject
                            text:(NSString *)text NS_DESIGNATED_INITIALIZER;
+
 - (instancetype)initWithFile:(NSString *)path
                     mimeType:(NSString *)mimeType NS_DESIGNATED_INITIALIZER;
+
+- (instancetype)initWithFile:(NSString *)path
+                    mimeType:(NSString *)mimeType
+                     subject:(NSString *)subject NS_DESIGNATED_INITIALIZER;
 
 - (instancetype)init
     __attribute__((unavailable("Use initWithSubject:text: instead")));
@@ -108,9 +114,23 @@ TopViewControllerForViewController(UIViewController *viewController) {
   return self;
 }
 
+- (instancetype)initWithFile:(NSString *)path
+                    mimeType:(NSString *)mimeType
+                     subject:(NSString *)subject {
+  self = [super init];
+  if (self) {
+    _path = path;
+    _mimeType = mimeType;
+    _subject = [subject isKindOfClass:NSNull.class] ? @"" : subject;
+  }
+  return self;
+}
+
 - (id)activityViewControllerPlaceholderItem:
     (UIActivityViewController *)activityViewController {
-  return @"";
+  return [self
+      activityViewController:activityViewController
+         itemForActivityType:@"dev.fluttercommunity.share_plus.placeholder"];
 }
 
 - (id)activityViewController:(UIActivityViewController *)activityViewController
@@ -119,13 +139,18 @@ TopViewControllerForViewController(UIViewController *viewController) {
     return _text;
   }
 
-  if ([_mimeType hasPrefix:@"image/"]) {
+  // If the shared file is an image return an UIImage for the placeholder
+  // to show a preview.
+  if ([activityType
+          isEqualToString:@"dev.fluttercommunity.share_plus.placeholder"] &&
+      [_mimeType hasPrefix:@"image/"]) {
     UIImage *image = [UIImage imageWithContentsOfFile:_path];
     return image;
-  } else {
-    NSURL *url = [NSURL fileURLWithPath:_path];
-    return url;
   }
+
+  // Return an NSURL for the real share to conserve the file name
+  NSURL *url = [NSURL fileURLWithPath:_path];
+  return url;
 }
 
 - (NSString *)activityViewController:
@@ -158,11 +183,42 @@ TopViewControllerForViewController(UIViewController *viewController) {
     (UIActivityViewController *)activityViewController
     API_AVAILABLE(macos(10.15), ios(13.0), watchos(6.0)) {
   LPLinkMetadata *metadata = [[LPLinkMetadata alloc] init];
-  if ([_text length] > 0) {
-    metadata.title = _text;
-  } else if ([_subject length] > 0) {
+
+  if ([_subject length] > 0) {
     metadata.title = _subject;
+  } else if ([_text length] > 0) {
+    metadata.title = _text;
   }
+
+  if (_path) {
+    NSString *extesnion = [_path pathExtension];
+
+    unsigned long long rawSize = (
+        [[[NSFileManager defaultManager] attributesOfItemAtPath:_path
+                                                          error:nil] fileSize]);
+    NSString *readableSize = [NSByteCountFormatter
+        stringFromByteCount:rawSize
+                 countStyle:NSByteCountFormatterCountStyleFile];
+
+    NSString *description = @"";
+
+    if (![extesnion isEqualToString:@""]) {
+      description =
+          [description stringByAppendingString:[extesnion uppercaseString]];
+      description = [description stringByAppendingString:@" • "];
+      description = [description stringByAppendingString:readableSize];
+    } else {
+      description = [description stringByAppendingString:readableSize];
+    }
+
+    // https://stackoverflow.com/questions/60563773/ios-13-share-sheet-changing-subtitle-item-description
+    metadata.originalURL = [NSURL fileURLWithPath:description];
+    if (_mimeType && [_mimeType hasPrefix:@"image/"]) {
+      metadata.imageProvider = [[NSItemProvider alloc]
+          initWithObject:[UIImage imageWithContentsOfFile:_path]];
+    }
+  }
+
   return metadata;
 }
 
@@ -312,26 +368,19 @@ TopViewControllerForViewController(UIViewController *viewController) {
           toResult:(FlutterResult)result {
   NSMutableArray *items = [[NSMutableArray alloc] init];
 
-  if (text || subject) {
-    [items addObject:[[SharePlusData alloc] initWithSubject:subject text:text]];
-  }
-
   for (int i = 0; i < [paths count]; i++) {
     NSString *path = paths[i];
-    NSString *pathExtension = [path pathExtension];
     NSString *mimeType = mimeTypes[i];
-    if ([pathExtension.lowercaseString isEqualToString:@"jpg"] ||
-        [pathExtension.lowercaseString isEqualToString:@"jpeg"] ||
-        [pathExtension.lowercaseString isEqualToString:@"png"] ||
-        [mimeType.lowercaseString isEqualToString:@"image/jpg"] ||
-        [mimeType.lowercaseString isEqualToString:@"image/jpeg"] ||
-        [mimeType.lowercaseString isEqualToString:@"image/png"]) {
-      UIImage *image = [UIImage imageWithContentsOfFile:path];
-      [items addObject:image];
-    } else {
-      NSURL *fileUrl = [NSURL fileURLWithPath:path];
-      [items addObject:fileUrl];
+    [items addObject:[[SharePlusData alloc] initWithFile:path
+                                                mimeType:mimeType
+                                                 subject:subject]];
+  }
+  if (text != nil) {
+    NSObject *data = [[NSURL alloc] initWithString:text];
+    if (data == nil) {
+      data = [[SharePlusData alloc] initWithSubject:subject text:text];
     }
+    [items addObject:data];
   }
 
   [self share:items
